requirements.txt
appdirs==1.4.4
attrs==22.2.0
autopep8==2.0.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
cssselect==1.2.0
fake-useragent==1.1.1
flake8==6.0.0
idna==3.4
importlib-metadata==6.0.0
lxml==4.9.2
mccabe==0.7.0
parse==1.19.0
pycodestyle==2.10.0
pyee==8.2.2
pyflakes==3.0.1
pyppeteer==1.0.2
pyquery==2.0.0
python-dotenv==0.21.1
requests==2.28.2
requests-html==0.10.0
soupsieve==2.3.2.post1
todoist-api-python==2.0.2
tomli==2.0.1
tqdm==4.64.1
urllib3==1.26.14
w3lib==2.1.1
websockets==10.4
zipp==3.12.0
